// src/services/rest-api/customer-service/model/bank-info-response-data-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Customer Service API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.267
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { BankInfoResponse } from './bank-info-response';

/**
 * 
 * @export
 * @interface BankInfoResponseDataResponse
 */
export interface BankInfoResponseDataResponse {
    /**
     * 
     * @type {number}
     * @memberof BankInfoResponseDataResponse
     */
    'status'?: number;
    /**
     * 
     * @type {{ [key: string]: Array<string>; }}
     * @memberof BankInfoResponseDataResponse
     */
    'errors'?: { [key: string]: Array<string>; } | null;
    /**
     * 
     * @type {string}
     * @memberof BankInfoResponseDataResponse
     */
    'message'?: string | null;
    /**
     * 
     * @type {BankInfoResponse}
     * @memberof BankInfoResponseDataResponse
     */
    'data'?: BankInfoResponse | null;
}